fix errors, peer dependencies
