+++ packages/queries/src/context.ts
@@ -1,13 +1,11 @@
 import { ethers } from "ethers";
 
 import type { SynthetixData } from '@synthetixio/data/build/node/src/types';
-import { NetworkId } from "@synthetixio/contracts-interface";
-
-type SynthetixJs = any;
+import { SynthetixJS, NetworkId } from "@synthetixio/contracts-interface";
 
 export interface QueryContext {
     networkId: NetworkId|null;
     provider: ethers.providers.Provider|null;
     snxData: SynthetixData|null,
-    snxjs: SynthetixJs|null
+    snxjs: SynthetixJS|null
 }
