--- packages/queries/src/modules.d.ts
@@ -1,1 +1,0 @@
-declare module 'synthetix-js';
