+++ packages/queries/src/types.ts
@@ -1,0 +1,69 @@
+import Wei from "@synthetixio/wei";
+import { CurrencyKey } from "./currency";
+
+export type GasPrices = {
+	fastest: number;
+	fast: number;
+	average: number;
+};
+
+export type GasSpeed = keyof GasPrices;
+
+export const GAS_SPEEDS: GasSpeed[] = ['average', 'fast', 'fastest'];
+
+export type BaseRateUpdate = {
+	timestamp: number;
+	rate: number;
+};
+
+export type BaseRateUpdates = BaseRateUpdate[];
+
+export type HistoricalRatesUpdates = {
+	rates: BaseRateUpdate[];
+	low: number;
+	high: number;
+	change: number;
+};
+
+export type MarketCap = {
+	marketCap: number;
+};
+
+export type Rates = Record<CurrencyKey, number>;
+
+export type HistoricalVolume = Record<CurrencyKey, Wei>;
+
+export type FrozenSynths = Set<CurrencyKey>;
+
+export type SynthSuspensionReason =
+	| 'system-upgrade'
+	| 'market-closure'
+	| 'circuit-breaker'
+	| 'emergency';
+
+export type SynthSuspended = {
+	isSuspended: boolean;
+	reasonCode: number;
+	reason: SynthSuspensionReason | null;
+};
+export type SynthBalance = {
+	currencyKey: CurrencyKey;
+	balance: Wei;
+	usdBalance: Wei;
+};
+
+export type SynthBalancesMap = Record<CurrencyKey, SynthBalance>;
+
+export type Balances = {
+	balancesMap: SynthBalancesMap;
+	balances: SynthBalance[];
+	totalUSDBalance: Wei;
+};
+
+export type TokenBalances = Record<
+	CurrencyKey,
+	{
+		balance: Wei;
+		token: any;
+	}
+>;
